Remove redundant proxy auth gate; keep only the /register OAuth rewrite (#984)
Instead of reviving the proxy's cookie-presence auth gate, remove it. It was a
weaker duplicate of the real gate: the server-side layout guards
((app)/layout.tsx via isAuthenticated(), complete-signup/layout.tsx) validate
the actual session (catching expired/revoked/forged cookies, which a
presence-only check cannot) on every dynamic render, backed by per-request
tRPC/API checks. Every protected route already self-guards there, so dropping
the proxy gate loses no coverage — it just removes a second, misleading auth
layer and consolidates auth into one place.

The proxy now does only the claude.ai OAuth workaround (method-split /register
-> /oauth/register rewrite), and config.matcher is scoped to ["/register"] so
middleware no longer runs on every request.

Replace proxy-public-paths.test.ts with proxy.test.ts covering the rewrite
behavior and that no auth gating happens.
